WIP Resumable upload
Added support for simple-uploader
